+++ gameplay/player/states/player_movement_state.gd
@@ -1,0 +1,6 @@
+class_name PlayerMovementState
+extends RefCounted
+
+
+func tick(_motor: PlayerMotor, _frame: InputFrame, _contacts: MovementContacts) -> void:
+	pass